+++ src/views/Login.vue
@@ -1,83 +1,99 @@
 <template>
-  <div class="wrapper">
+  <div class="main">
     <div class="login">
       <h1 class="title-1">{{ $t('login.login') }}</h1>
-      <div class="sections">
-        <div class="section">
-          <div v-if="loginError === 1" class="error">
-            {{ $t('login.error.notRegistered') }}
+      <div v-if="loginLoading && loginError === null">
+        <loading-small></loading-small>
+        {{ $t('login.loading') }}
+      </div>
+      <div v-else class="section">
+        <div v-if="loginError === 1" class="error">
+          {{ $t('login.error.notRegistered') }}
 
-            <router-link :to="{ name: 'request-access' }">{{ $t('login.requestAccess') }} </router-link>
-          </div>
-
-          <div v-if="loginError === 2" class="error">
-            {{ $t('login.error.googleError') }}
-          </div>
-        </div>
-        <div v-if="showForm" class="login__form">
-          <div v-if="loginError === 3" class="error">{{ $t('login.error.wrongPassword') }}</div>
-
-          <div v-if="loginError === 4" class="error">
-            {{ $t('login.error.userNotFound') }}
-          </div>
-          <validation-observer v-slot="{ handleSubmit }">
-            <form id="login" @submit.prevent="handleSubmit(loginWithEmail)">
-              <form-component
-                v-model="email"
-                :label="$t('login.email')"
-                input-type="input"
-                name="email"
-                rules="required|email"
-                type="email"
-                data-cy="login-username-input"
-              />
-
-              <form-component
-                v-model="password"
-                :label="$t('login.password')"
-                input-type="input"
-                name="password"
-                rules="required"
-                type="password"
-                data-cy="login-username-password"
-              />
-            </form>
-          </validation-observer>
-          <button class="btn btn--pri" form="login">{{ $t('login.login') }}</button>
+          <router-link :to="{ name: 'request-access' }">{{ $t('login.requestAccess') }} </router-link>
         </div>
 
-        <div class="login__footer">
-          <button class="btn btn--icon btn--pri" @click="loginWithGoogle">
-            <i class="icon fab fa-fw fa-google" />
-            {{ $t('login.google') }}
-          </button>
-
-          <div class="login__secondary">
-            <button class="btn btn--ghost" data-cy="login-username" @click="showForm = true">
-              {{ $t('login.loginWithUsername') }}
-            </button>
-            <router-link class="btn btn--ghost" :to="{ name: 'request-access' }" data-cy="login-request">
-              {{ $t('login.requestAccess') }}
-            </router-link>
-          </div>
+        <div v-if="loginError === 2" class="error">
+          {{ $t('login.error.googleError') }}
         </div>
+      </div>
+      <div v-if="showForm" class="login__form">
+        <div v-if="loginError === 3" class="error">{{ $t('login.error.wrongPassword') }}</div>
+
+        <div v-if="loginError === 4" class="error">
+          {{ $t('login.error.userNotFound') }}
+        </div>
+        <validation-observer v-slot="{ handleSubmit }">
+          <form id="login" @submit.prevent="handleSubmit(loginWithEmail)">
+            <form-component
+              v-model="email"
+              :label="$t('login.email')"
+              input-type="input"
+              name="email"
+              rules="required|email"
+              type="email"
+              data-cy="login-username-input"
+            />
+
+            <form-component
+              v-model="password"
+              :label="$t('login.password')"
+              input-type="input"
+              name="password"
+              rules="required"
+              type="password"
+              data-cy="login-username-password"
+            />
+          </form>
+        </validation-observer>
+        <button class="btn btn--pri" form="login">{{ $t('login.login') }}</button>
+      </div>
+
+      <div v-if="!loginLoading || loginError !== null" class="login__footer">
+        <button v-if="providers.includes('google')" class="btn btn--icon btn--pri" @click="loginWithGoogle">
+          <i class="icon fab fa-fw fa-google" />
+          {{ $t('login.google') }}
+        </button>
+
+        <button
+          v-if="providers.includes('keycloak')"
+          class="btn btn--pri"
+          data-cy="login-username"
+          @click="loginWithKeycloak"
+        >
+          {{ $t('login.keycloak') }}
+        </button>
+
+        <button
+          v-if="providers.includes('email')"
+          class="btn btn--ghost"
+          data-cy="login-username"
+          @click="showForm = true"
+        >
+          {{ $t('login.loginWithUsername') }}
+        </button>
+        <router-link class="btn btn--ghost" :to="{ name: 'request-access' }" data-cy="login-request">
+          {{ $t('login.requestAccess') }}
+        </router-link>
       </div>
     </div>
   </div>
 </template>
 
 <script>
-import { mapMutations, mapState } from 'vuex';
-import { auth, loginProvider } from '@/config/firebaseConfig';
+import { mapState, mapActions } from 'vuex';
+import { auth, functions, loginProvider } from '@/config/firebaseConfig';
 import i18n from '@/locale/i18n';
+import LoadingSmall from '@/components/LoadingSmall.vue';
 
 export default {
   name: 'Login',
 
+  components: { LoadingSmall },
+
   data: () => ({
     email: '',
     password: '',
-    pending: false,
     showForm: false,
   }),
 
@@ -88,34 +104,69 @@ export default {
   },
 
   computed: {
-    ...mapState(['user', 'loginError']),
+    ...mapState(['user', 'loginError', 'providers', 'keycloak', 'authenticated', 'loginLoading']),
+  },
+
+  watch: {
+    authenticated: {
+      immediate: true,
+      async handler() {
+        if (this.providers.includes('keycloak') && this.authenticated) {
+          await this.setLoginLoading(true);
+          await this.setLoginError(null);
+          try {
+            const myCall = functions.httpsCallable('createCustomToken');
+            const customToken = await myCall(this.keycloak.token);
+            await auth.signInWithCustomToken(customToken.data);
+            await this.setLoginLoading(false);
+            this.$toasted.show(this.$t('toaster.welcome', { user: this.keycloak.idTokenParsed.name }));
+          } catch (e) {
+            this.keycloak.logout({ redirectUri: `${process.env.VUE_APP_KEYCLOAK_ERROR_URL}${e.code}` });
+          }
+        }
+      },
+    },
   },
 
   methods: {
-    ...mapMutations(['SET_LOGIN_ERROR']),
-    async loginWithGoogle() {
-      this.pending = true;
+    ...mapActions(['initKeycloak', 'cleanKeycloak', 'setLoading', 'setLoginLoading', 'setLoginError']),
+
+    async loginWithKeycloak() {
+      await this.setLoginLoading(true);
+      await this.setLoginError(null);
       try {
-        const user = await auth.signInWithPopup(loginProvider);
-        this.$toasted.show(this.$t('toaster.welcome', { user: user.name ? user.name : '' }));
+        await this.keycloak.login();
       } catch (e) {
-        this.pending = false;
-        this.SET_LOGIN_ERROR(2);
+        throw new Error(e);
       }
+      await this.setLoginLoading(false);
+    },
+
+    async loginWithGoogle() {
+      await this.setLoginLoading(true);
+      await this.setLoginError(null);
+      try {
+        const { user } = await auth.signInWithPopup(loginProvider);
+        this.$toasted.show(this.$t('toaster.welcome', { user: user.displayName ? user.displayName : '' }));
+      } catch (e) {
+        await this.setLoginError(2);
+      }
+      await this.setLoginLoading(false);
     },
 
     async loginWithEmail() {
-      this.pending = true;
+      await this.setLoginLoading(true);
+      await this.setLoginError(null);
 
       try {
         await auth.signInWithEmailAndPassword(this.email, this.password);
       } catch (err) {
         console.log(err);
-        this.pending = false;
+        await this.setLoginLoading(false);
         if (err.code === 'auth/wrong-password') {
-          this.SET_LOGIN_ERROR(3);
+          await this.setLoginError(3);
         } else if (err.code === 'auth/user-not-found') {
-          this.SET_LOGIN_ERROR(4);
+          await this.setLoginError(4);
         }
       }
     },
@@ -129,32 +180,18 @@ export default {
 .login {
   display: flex;
   flex-direction: column;
-  width: span(10);
-  margin-left: span(1, 1);
+  width: span(12);
   padding: 2rem;
   background: white;
   border-radius: 3px;
   box-shadow: 0 2px 4px rgba($color-grey-400, 0.3);
 
-  @media screen and (min-width: bp(xs)) {
-    width: span(8);
-    margin-top: 2rem;
-    margin-left: span(2, 1);
+  @media screen and (min-width: bp(m)) {
+    width: span(4);
   }
 
   @media screen and (min-width: bp(s)) {
     width: span(6);
-    margin-top: 3rem;
-    margin-left: 0;
-  }
-
-  @media screen and (min-width: bp(m)) {
-    width: span(5, 0, span(9));
-    margin-top: 5rem;
-  }
-
-  @media screen and (min-width: bp(l)) {
-    width: span(4, 0, span(10));
   }
 }
 
@@ -164,6 +201,9 @@ export default {
 }
 
 .login__footer {
+  display: flex;
+  flex-direction: column;
+  gap: 0.5rem 0.5rem;
   margin-top: 2rem;
 }